ReaderMode(invent.ids remaded) , refractoring

+++ requestwidget.cpp
@@ -26,6 +26,8 @@ void RequestWidget::loadTable()
     QSqlQueryModel* requests_model {new QSqlQueryModel};
     requests_model->setQuery("SELECT * FROM request");
     ui->tableView->setModel(requests_model);
+
+    ui->tableView->verticalHeader()->setVisible(false); // remove indexes on leftside
     ui->tableView->setMinimumWidth(ui->tableView->columnWidth(0)
                                  * requests_model->columnCount()
                                  + 10);
@@ -33,8 +35,11 @@ void RequestWidget::loadTable()
 
 void RequestWidget::on_tableView_activated(const QModelIndex &index)
 {
-    const QString request_id {ui->tableView->model()->index(index.row(), 0).data().toString()};
-    ui->idLabel->setText(request_id);
+    // Set request_id as idLabel text
+    ui->idLabel->setText(
+                ui->tableView->model()->index(index.row(), 0).data().toString());
+
+    ui->acceptPushButton->setEnabled(true);
 }
 
 void RequestWidget::on_acceptPushButton_clicked()
@@ -48,20 +53,21 @@ void RequestWidget::on_acceptPushButton_clicked()
     QString inventoryed_id {query.value(0).toString()};
     QString reader_id      {query.value(1).toString()};
 
-    QSqlQuery query2;
-    query2.exec("INSERT INTO return_table (inventoryed_id, issue_date, return_date, reader_id) "
-                "VALUES (" + inventoryed_id
-                           + ", CURDATE(), "
-                           + "CURDATE() + INTERVAL "
-                           + QString::number(ui->timeSpinBox->value()) + " "
-                           + ui->periodComboBox->currentText() + ", "
-                           + reader_id + ")");
-
-    QSqlQuery query3;
-    query3.exec("DELETE FROM request "
-                "WHERE request_id = " + request_id);
-
-    QMessageBox::information(this, "Success", "Successfully accepted.");
-
-    loadTable();
+    QSqlQuery query2{"INSERT INTO return_table (inventoryed_id, issue_date, return_date, reader_id) "
+                     "VALUES (" + inventoryed_id
+                                + ", CURDATE(), "
+                                + "CURDATE() + INTERVAL "
+                                + QString::number(ui->timeSpinBox->value()) + " "
+                                + ui->periodComboBox->currentText() + ", "
+                                + reader_id + ")"};
+    if (query2.exec()) {
+        QMessageBox::information(this, "Success", "Successfully accepted.");
+        QSqlQuery query3;
+        query3.exec("DELETE FROM request "
+                    "WHERE request_id = " + request_id);
+        loadTable(); // update after erasing.
+        ui->acceptPushButton->setEnabled(false); // id is not choosed, so disable button
+    } else {
+        QMessageBox::information(this, "Success", "Successfully accepted.");
+    }
 }
